requirements.txt
streamlit>=1.24.0
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.2.0
openai>=1.0.0
supabase>=1.0.3
python-dotenv>=1.0.0
requests>=2.28.0
nomic>=2.0.3
# Document processing libraries
PyPDF2>=3.0.0
python-docx>=0.8.11
